# packages/api-client/openapi.yaml
openapi: "3.1.0"
info:
  title: Project Anvil API
  description: Contract-first API specification for the Anvil federated Alphabet ecosystem
  version: "0.1.0"
  contact:
    name: Project Anvil
    url: https://github.com/yethikrishna/project-anvil
servers:
  - url: http://localhost:3100
    description: Drive API
  - url: http://localhost:3102
    description: Docs API
  - url: http://localhost:3300
    description: YouTube API
  - url: http://localhost:3400
    description: Maps API
  - url: http://localhost:3500
    description: Search API
  - url: http://localhost:3600
    description: Gmail API

# ── Common Components ───────────────────────────────────
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Keycloak OIDC access token
  parameters:
...
    docIdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
...
    Document:
      type: object
      required: [id, title, ownerId, collaborators, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        ownerId:
          type: string
        title:
          type: string
        content:
          type: string
          description: HTML content
        version:
          type: integer
        collaborators:
          type: array
          items:
            $ref: "#/components/schemas/DocumentCollaborator"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    DocumentCollaborator:
      type: object
      required: [userId, role]
      properties:
        userId:
          type: string
        role:
          type: string
          enum: [owner, editor, viewer]

    # YouTube
...
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
...
security:
  - BearerAuth: []

# ── Paths ──────────────────────────────────────────────
paths:
...
  /api/documents:
    get:
      tags: [Docs]
      operationId: listDocuments
      summary: List user's documents
      responses:
        "200":
          description: Document list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"

    post:
      tags: [Docs]
      operationId: createDocument
      summary: Create a new document
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
      responses:
        "201":
          description: Document created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
  /api/documents/{id}:
    get:
      tags: [Docs]
      operationId: getDocument
      summary: Get a document
      parameters:
        - $ref: "#/components/parameters/docIdParam"
      responses:
        "200":
          description: Document data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "404":
          $ref: "#/components/responses/NotFound"

    patch:
      tags: [Docs]
      operationId: updateDocument
      summary: Update document (auto-save)
      parameters:
        - $ref: "#/components/parameters/docIdParam"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                content:
                  type: string
                ydocState:
                  type: string
      responses:
        "200":
          description: Updated
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      tags: [Docs]
      operationId: deleteDocument
      summary: Delete a document
      parameters:
        - $ref: "#/components/parameters/docIdParam"
      responses:
        "200":
          description: Deleted
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/documents/search:
    get:
      tags: [Docs]
      operationId: searchDocuments
      summary: Search documents by title
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"

  # ── Search ───────────────────────────────────────────
...